Split languageDetector into ParserLanguageDetector and SqlDialectDetector
